move webshell related logic to own module
